# Perf/loop over kids first