--- a/packages/treetime/src/commands/ancestral/fitch.rs
+++ b/packages/treetime/src/commands/ancestral/fitch.rs
@@ -1,7 +1,4 @@
 #![allow(dead_code)]
-
-use std::collections::BTreeMap;
-
 use crate::alphabet::alphabet::{FILL_CHAR, NON_CHAR, VARIABLE_CHAR};
 use crate::graph::breadth_first::GraphTraversalContinuation;
 use crate::io::fasta::FastaRecord;
@@ -9,7 +6,7 @@ use crate::representation::graph_sparse::{
   Deletion, ParsimonySeqDis, SparseGraph, SparseNode, SparseSeqDis, SparseSeqEdge, SparseSeqInfo, SparseSeqNode,
 };
 use crate::representation::partitions_parsimony::{PartitionParsimony, PartitionParsimonyWithAln};
-use crate::representation::state_set::{BitSet128, StateSet, StateSetStatus};
+use crate::representation::state_set::{StateSet, StateSetStatus};
 use crate::seq::composition::Composition;
 use crate::seq::indel::InDel;
 use crate::seq::mutation::Sub;
@@ -167,60 +164,33 @@ fn fitch_backwards(graph: &SparseGraph, sparse_partitions: &[PartitionParsimony]
         }
       }
 
+      // // Process all positions where the children are fixed or completely unknown in some children.
+      // loop over all children
       for &(child, _) in &children {
         for (pos, parent_state) in sequence.iter_mut().enumerate() {
           let cstate = child.sequence[pos];
-          if *parent_state==cstate || *parent_state==NON_CHAR {
-            continue;
+          if *parent_state == cstate || *parent_state == NON_CHAR {
+            continue; // if parent is equal to child state or we know its a non-char, skip
           }
           if alphabet.is_canonical(cstate) {
-            if *parent_state==FILL_CHAR {
-              *parent_state=cstate;
-            } else{
+            if *parent_state == FILL_CHAR { // if cstate is canonical and parent is still FILL_STATE, set parent_state
+              *parent_state = cstate;
+            } else { // otherwise set or update the variable state
               if seq_dis.variable.contains_key(&pos) {
-                seq_dis.variable.insert(pos, StateSet::from_union(&[seq_dis.variable[&pos], StateSet::from_char(cstate)]));
+                seq_dis.variable.insert(
+                  pos,
+                  StateSet::from_union(&[seq_dis.variable[&pos], StateSet::from_char(cstate)]),
+                );
               } else {
-                seq_dis.variable.insert(pos,  StateSet::from_slice(&[*parent_state, cstate]));
+                seq_dis
+                  .variable
+                  .insert(pos, StateSet::from_slice(&[*parent_state, cstate]));
               }
               *parent_state = VARIABLE_CHAR;
             }
           }
         }
       }
-
-      // // Process all positions where the children are fixed or completely unknown in some children.
-      // for (pos, parent_state) in sequence.iter_mut().enumerate() {
-      //   if *parent_state != FILL_CHAR {
-      //     continue;
-      //   }
-
-      //   let mut determined_states = Vec::with_capacity(alphabet.n_chars());
-      //   for &(child, _) in &children {
-      //     let state = child.sequence[pos];
-      //     if !determined_states.contains(&state) && alphabet.is_canonical(state) {
-      //       determined_states.push(state);
-      //     }
-      //   }
-
-      //   // Find the state of the current node at this position
-      //   *parent_state = match determined_states.as_slice() {
-      //     [state] => {
-      //       // All children have the same state, that will be the state of the current node
-      //       *state
-      //     }
-      //     [] => {
-      //       // No child states. Impossible
-      //       unreachable!("No child states. This is impossible");
-      //     }
-      //     states => {
-      //       // Child states differ. This is variable state.
-      //       // Save child states and postpone the decision until forward pass.
-      //       let dis = states.iter().collect();
-      //       seq_dis.variable.insert(pos, dis);
-      //       VARIABLE_CHAR
-      //     }
-      //   };
-      // }
 
       // Process insertions and deletions.