--- a/packages/treetime/src/representation/mod.rs
+++ b/packages/treetime/src/representation/mod.rs
@@ -1,7 +1,7 @@
+pub mod bitset128;
 pub mod graph_dense;
 pub mod graph_sparse;
 pub mod infer_dense;
 pub mod partitions_likelihood;
 pub mod partitions_parsimony;
 pub mod state_set;
-pub mod bitset128;