--- a/packages/treetime/src/utils/mod.rs
+++ b/packages/treetime/src/utils/mod.rs
@@ -1,5 +1,4 @@
 pub mod assert;
-pub mod openblas;
 pub mod container;
 pub mod datetime;
 pub mod error;
@@ -8,6 +7,7 @@ pub mod global_init;
 pub mod interval;
 pub mod manyzip;
 pub mod ndarray;
+pub mod openblas;
 pub mod random;
 pub mod string;
 pub mod vec;