small experiment: this is a bit faster for me
```
richard@X1-2024:~/Projects_GitHub/TreeTime/treetime_rs$ hyperfine --warmup 1 --show-output 'target/release/treetime ancestral --method-anc=parsimony --dense=false --tree=data/mpox/clade-ii/500/tree.nwk --outdir=
tmp/smoke-tests/ancestral/marginal/mpox/clade-ii/500 data/mpox/clade-ii/500/aln.fasta.xz'                
Benchmark 1: target/release/treetime ancestral --method-anc=parsimony --dense=false --tree=data/mpox/clade-ii/500/tree.nwk --outdir=tmp/smoke-tests/ancestral/marginal/mpox/clade-ii/500 data/mpox/clade-ii/500/aln
.fasta.xz                                                                                                                                                                                                          
  Time (mean ± σ):      1.016 s ±  0.025 s    [User: 2.100 s, System: 2.086 s]                                                                                                                                     
  Range (min … max):    0.980 s …  1.064 s    10 runs                                                                                                                                                              
```

vs

```
richard@X1-2024:~/Projects_GitHub/TreeTime/treetime_rs$ hyperfine --warmup 1 --show-output 'target/release/treetime ancestral --method-anc=parsimony --dense=false --tree=data/mpox/clade-ii/500/tree.nwk --outdir=
tmp/smoke-tests/ancestral/marginal/mpox/clade-ii/500 data/mpox/clade-ii/500/aln.fasta.xz'                
Benchmark 1: target/release/treetime ancestral --method-anc=parsimony --dense=false --tree=data/mpox/clade-ii/500/tree.nwk --outdir=tmp/smoke-tests/ancestral/marginal/mpox/clade-ii/500 data/mpox/clade-ii/500/aln
.fasta.xz                                                                                                
  Time (mean ± σ):      1.157 s ±  0.027 s    [User: 3.124 s, System: 2.120 s]                           
  Range (min … max):    1.124 s …  1.198 s    10 runs                                                    
```

not really worth the trouble unless you feel it is cleaner...